Move DashboardNavbar components to components/common directory
