+++ frontend/src/components/common/DashboardNavbar/DashboardNavbar.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import "../../../styles/components/dashboard/dashboard-navbar.scss";
-import ParentLink from "./navbar/ParentLink";
-import AccountLinks from "./navbar/AccountLinks";
-import { dashboardNavLinks } from "../../../../constants/navLinksConfig";
+import { dashboardNavLinks } from "../../../constants/navLinksConfig";
+import ParentLink from "./components/ParentLink";
+import AccountLinks from "./components/AccountLinks";
 
 function DashboardNavbar() {
 	const [dashboardNavbarExpanded, setDashboardNavbarExpanded] =
